Remove namespace std using directives

+++ smips/algorithms/decomposition/ald/is_integer.cpp
@@ -3,6 +3,6 @@
 bool Ald::is_integer(double val)
 {
     float intpart;
-    float frac = modf(val, &intpart);
+    float frac = std::modf(static_cast<float>(val), &intpart);
     return frac == 0 || frac == 1;
 }
