+++ smips/algorithms/decomposition/zk/is_integer.cpp
@@ -3,7 +3,6 @@
 bool ZK::is_integer(double val, double precision)
 {
     float intpart;
-    float frac = modf(val, &intpart);
+    float frac = modf(static_cast<float>(val), &intpart);
     return frac < precision || frac > 1 - precision;
-    // return frac == 0 || frac == 1;
 }
